+++ tools/schemas/schema.ts
@@ -1,12 +1,14 @@
 import { z } from 'zod';
 
+const UrlSchema = z.record(
+  z.string(),
+  z.union([z.string(), z.array(z.string())]),
+);
+
 const MonorepoSchema = z.object({
-  repoGroups: z.record(z.string(), z.union([z.string(), z.array(z.string())])),
-  orgGroups: z.record(z.string(), z.union([z.string(), z.array(z.string())])),
-  patternGroups: z.record(
-    z.string(),
-    z.union([z.string(), z.array(z.string())]),
-  ),
+  repoGroups: UrlSchema,
+  orgGroups: UrlSchema,
+  patternGroups: UrlSchema,
 });
 
 export { MonorepoSchema };
